+++ src/main/java/nextstep/ladder/domain/Height.java
@@ -1,7 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.Comparator;
 import java.util.Objects;
+
+import nextstep.ladder.exception.InvalidHeightException;
 
 public class Height implements Comparator<Height> {
     private static final int MIN_HEIGHT_LENGTH = 1;
@@ -14,7 +16,7 @@ public class Height implements Comparator<Height> {
 
     private void validate(final int height) {
         if (height < MIN_HEIGHT_LENGTH) {
-            throw new InvalidHeightException("사다리 높이는 최대 " + MIN_HEIGHT_LENGTH +"이상 이어야 합니다.");
+            throw new InvalidHeightException("사다리 높이는 최대 " + MIN_HEIGHT_LENGTH + "이상 이어야 합니다.");
         }
     }
