+++ src/main/java/nextstep/ladder/exception/InvalidHeightException.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.exception;
 
 public class InvalidHeightException extends LadderException {
     public InvalidHeightException(final String message) {
