+++ src/main/java/nextstep/ladder/exception/InvalidPlayerNameException.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.exception;
 
 public class InvalidPlayerNameException extends LadderException {
     public InvalidPlayerNameException(final String message) {
