+++ src/main/java/nextstep/ladder/exception/LadderException.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.exception;
 
 public class LadderException extends RuntimeException {
     public LadderException(final String message) {
