+++ src/main/java/nextstep/ladder/exception/LineException.java
@@ -1,4 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.exception;
+
+import nextstep.ladder.exception.LadderException;
 
 public class LineException extends LadderException {
     public LineException(final String message) {
